Added chan2 check for extraction tests. Removed extraction utils and moved to detection utils. source and sparsery now dont depend on extraction module.
